cms_demo/requirements.txt
asgiref==3.2.10
backcall==0.2.0
cffi==1.14.2
colorama==0.4.3
cryptography==3.0
decorator==4.4.2
Django==3.1
django-extensions==3.0.5
django-rest-framework==0.1.0
djangorestframework==3.11.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.6
pycparser==2.20
Pygments==2.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
